Tatduc

+++ app/src/main/java/com/example/appquanlynhansu/Model/ChucVu.kt
@@ -2,5 +2,5 @@ package com.example.appquanlynhansu.Model
 
 data class ChucVu(
     val tenchucvu : String,
-    val luongcoban: Double,
+    val luongcoban: Long,
 )
